replace Java String contains method with indexOf method
contains works on JS String in Nashorn, but it throws an error in GraalVM's JS engine.
This leads to runtime errors during tests and in SQLDev and SQLcl if GraalVMs JS engine is used.
indexOf is available for a Java and JS String.
Therefore this change improves the compatibility between Nashorn other JS engines.
For the time being Nashorn must be supported, since SQLDev and SQcl uses it when it is available.